Module Tezos_client_009_PsFLoren.Annotated_manager_operation

Annotated manager operations are wrappers used to accumulate information (especially about limits) on the operation prior to the injection.

type _ t =

| Manager_info : {

source : Tezos_protocol_009_PsFLoren.Protocol.Alpha_context.public_key_hash option;

fee : Tezos_protocol_009_PsFLoren.Protocol.Alpha_context.Tez.t Limit.t;

gas_limit : Tezos_protocol_009_PsFLoren.Protocol.Alpha_context.Gas.Arith.integral Limit.t;

storage_limit : Z.t Limit.t;

counter : Z.t option;

operation : 'kind Tezos_protocol_009_PsFLoren.Protocol.Alpha_context.manager_operation;

} -> 'kind t

type packed =

| Annotated_manager_operation : 'kind t -> packed

type _ annotated_list =

| Single_manager : 'kind t -> 'kind annotated_list

| Cons_manager : 'kind t * 'rest annotated_list -> ('kind * 'rest) annotated_list

The annotated_list type helps making contents_list from a list of manager_operations. Its construction mimics contents_list in order to keep consistent types when calling inject_manager_operation and inject_operation.

type packed_annotated_list =

| Manager_list : 'kind annotated_list -> packed_annotated_list

val manager_to_list : packed_annotated_list -> packed list

Convert a list of annotated operations to a list of packed annotated operations

val manager_of_list : packed list -> packed_annotated_list

Converse of manager_to_list

val join_fee : Tezos_protocol_009_PsFLoren.Protocol.Alpha_context.Tez.t Limit.t -> 'a t -> ('a t, Tezos_error_monad.TzCore.error list) Stdlib.result

join_fee fee op updates op.fee to Limit.join op.fee fee and fails if the join fails

val set_fee : Tezos_protocol_009_PsFLoren.Protocol.Alpha_context.Tez.t Limit.t -> 'a t -> 'a t

set_fee fee op updates op.fee to fee

val join_gas_limit : Tezos_protocol_009_PsFLoren.Protocol.Alpha_context.Gas.Arith.integral Limit.t -> 'a t -> ('a t, Tezos_error_monad.TzCore.error list) Stdlib.result

See join_fee

val set_gas_limit : Tezos_protocol_009_PsFLoren.Protocol.Alpha_context.Gas.Arith.integral Limit.t -> 'a t -> 'a t

See set_fee

val join_storage_limit : Z.t Limit.t -> 'a t -> ('a t, Tezos_error_monad.TzCore.error list) Stdlib.result

See join_fee

val set_storage_limit : Z.t Limit.t -> 'a t -> 'a t

See set_fee

val set_counter : Tezos_protocol_009_PsFLoren.Protocol.Alpha_context.counter -> 'a t -> ('a t, Tezos_error_monad.TzCore.error list) Stdlib.result

Set the counter of the annotated operation. Fail if the counter is already set.

val set_source : Tezos_protocol_009_PsFLoren.Protocol.Alpha_context.public_key_hash -> 'a t -> ('a t, Tezos_error_monad.TzCore.error list) Stdlib.result

Set the source of the operation. Fail if the source is already set.

val manager_from_annotated : 'a t -> ('a Tezos_protocol_009_PsFLoren.Protocol.Alpha_context.Kind.manager Tezos_protocol_009_PsFLoren.Protocol.Alpha_context.contents, Tezos_error_monad.TzCore.error list) Stdlib.result

Convert an annotated manager operation to a proper manager operation. Fail if some fields in the annotated operation are not set.

val manager_list_from_annotated : 'kind annotated_list -> ('kind Tezos_protocol_009_PsFLoren.Protocol.Alpha_context.Kind.manager Tezos_protocol_009_PsFLoren.Protocol.Alpha_context.contents_list, Tezos_error_monad.TzCore.error list) Stdlib.result